Move structure to common
